requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==3.0.2
packaging==24.1
pillow==11.0.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.6
